pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "multilingual_watermarking"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Adam Gorgo\u0144" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "accelerate>=1.6.0",
    "bitsandbytes>=0.45.5",
    "huggingface-hub[cli]>=0.30.2",
    "ipython",
    "jupyterlab",
    "matplotlib",
    "nltk>=3.9.1",
    "notebook",
    "numpy",
    "openai>=1.91.0",
    "pandas",
    "pip",
    "protobuf>=6.30.2",
    "pytest",
    "python-dotenv",
    "pyyaml>=6.0.2",
    "ruff",
    "scikit-learn",
    "sentencepiece>=0.2.0",
    "spacy>=3.8.7",
    "torch>=2.7.0",
    "transformers>=4.51.3",
]
requires-python = "~=3.10.0"


[tool.ruff]
line-length = 99
src = ["multilingual_watermarking"]
include = ["pyproject.toml", "multilingual_watermarking/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["multilingual_watermarking"]
force-sort-within-sections = true

